Rename bot.py to agent.py
